requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
anndata>=0.8.0

# Machine Learning & Statistics
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Dimensionality Reduction & Clustering
umap-learn>=0.5.0
leidenalg>=0.8.0
python-igraph>=0.9.0

# Genomic Data
pyranges>=0.0.100

# Parallel & Performance
joblib>=1.1.0
tqdm>=4.62.0
